Realizado a Criação da Classe CarrinhoController e atributos metodos get/post/delete em ControllerProduct. criado classe de Exception

+++ src/main/java/com/example/supera/api/ProductModel/Product.java
@@ -19,7 +19,7 @@ public class Product {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	private String name;
-	private BigDecimal price;
+	private double price;
 	private short score;
 	private String image;
 
@@ -39,11 +39,11 @@ public class Product {
 		this.name = name;
 	}
 
-	public BigDecimal getPrice() {
+	public double getPrice() {
 		return price;
 	}
 
-	public void setPrice(BigDecimal price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
